+++ src/components/notification/NotificationEmpty.vue
@@ -1,0 +1,7 @@
+<!-- NotificationEmpty.vue -->
+<template>
+  <div class="flex flex-col items-center justify-center text-gray-400 py-10 gap-3">
+    <i class="fa fa-bell-slash text-3xl"></i>
+    <p class="text-sm">표시할 알림이 없습니다.</p>
+  </div>
+</template>
