+++ src/components/notification/NotificationList.vue
@@ -1,6 +1,6 @@
-<!-- /src/components/system/NotificationList.vue -->
+<!-- NotificationList.vue -->
 <template>
-  <div class="notification-list">
+  <div class="flex flex-col gap-3 mt-5">
     <template v-if="!loading && notifications.length">
       <NotificationCard
         v-for="notification in notifications"
@@ -10,32 +10,12 @@
       />
     </template>
     <NotificationEmpty v-else-if="!loading && !notifications.length" />
-    <div v-else class="loading">불러오는 중...</div>
+    <div v-else class="text-center text-gray-500 text-sm mt-10">불러오는 중...</div>
   </div>
 </template>
 
 <script setup>
 import NotificationCard from './NotificationCard.vue'
 import NotificationEmpty from './NotificationEmpty.vue'
-
-defineProps({
-  notifications: Array,
-  loading: Boolean,
-})
+defineProps({ notifications: Array, loading: Boolean })
 </script>
-
-<style scoped>
-.notification-list {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  margin-top: 20px;
-}
-
-.loading {
-  text-align: center;
-  color: #6c757d;
-  font-size: 14px;
-  margin-top: 40px;
-}
-</style>
